requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
jws==0.1.3
oauth2client==3.0.0
Pillow==8.4.0
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.6
pyserial==3.5
python-firebase==1.2.1
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
